Use display grid to fix padding issues
I initially used flex displays but it turns out to be a lot of nested
components, then I learned that using grid was the way to go and used
that in the main container and child cards.

I also added the date in the left-side of the card to know when it is
posted without needing to go to the post itself.