🚀 Migrate to SvelteKit (#69)
after a long and tedious process of 125 commits, it's finally here

there should be more that can be improved here, and it is by no
means done, especially with SvelteKit still in beta. but, it's better
to get this on master before doing any more improvements. the
content has been blocked from update for a long time because
of this, and it'll be so much better now that it has been migrated.

external build process needs to be configured again as it's still
configured for Sapper, other than that, this should be all there is.